% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm_cons.R
\name{alm_gd}
\alias{alm_gd}
\title{gradient function in the Augmented Lagrangian method}
\usage{
alm_gd(x, J, G, S, gamma, rho, Pair)
}
\arguments{
\item{x}{input free parameters in bi-factor model}

\item{J}{the number of items}

\item{G}{the number of group factors}

\item{S}{the sample covariance matrix}

\item{gamma}{a matrix-like ALM parameter}

\item{rho}{a scalar-like ALM parameter}

\item{Pair}{a list containing all pairs of columns in the equality constraints}
}
\value{
the gradient of the objective function in the Augmented Lagrangian method
}
\description{
gradient function in the Augmented Lagrangian method
}
