% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_step.R
\name{alm_solve}
\alias{alm_solve}
\title{The Augmented Lagrangian method for exploring an exact bi-factor model}
\usage{
alm_solve(
  J,
  G,
  S,
  Pair,
  gamma_0,
  rho_0,
  rho_sigma,
  theta,
  tol,
  max_iter,
  x_init = numeric(0)
)
}
\arguments{
\item{J}{the number of items}

\item{G}{the number of group factors}

\item{S}{the sample covariance matrix}

\item{Pair}{the constraint for bi-factor model}

\item{gamma_0}{the initial value for gamma, a matrix-like ALM parameter}

\item{rho_0}{the initial value for rho, a scalar, ALM parameter}

\item{rho_sigma}{a hyper-parameter for ALM method}

\item{theta}{a hyper-parameter for ALM method}

\item{tol}{the tolerance for ALM method}

\item{max_iter}{the maximal iteration number of the ALM method}

\item{x_init}{initial value for parameters}
}
\value{
A list containing:
\itemize{
\item \code{x_est}: the estimated parameters of the bi-factor model
\item \code{iter_num}: the iteration number of the bi-factor model
\item \code{L_est}: the estimated loading matrix
\item \code{Psi_est}: the estimated correlation matrix
\item \code{D_est}: the estimated unique variance matrix
\item \code{Q_est}: the estimated bi-factor structure
}
}
\description{
The Augmented Lagrangian method for exploring an exact bi-factor model
}
