% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{Average_correct}
\alias{Average_correct}
\title{Average Correctness Criterion for bi-factor structure}
\usage{
Average_correct(Q, Q_est, Rot_list = list())
}
\arguments{
\item{Q}{the default bi-factor structure}

\item{Q_est}{the estimated bi-factor structure}

\item{Rot_list}{the rotation matrix considered in the criterion, default is an empty list}
}
\value{
value ranging between 0-1, a larger value indicates a higher correctness
}
\description{
Average Correctness Criterion for bi-factor structure
}
