% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bf_compress}
\alias{bf_compress}
\title{compress a bi-factor structure from loading}
\usage{
bf_compress(L, tol)
}
\arguments{
\item{L}{the input loading matrix}

\item{tol}{the tolerance of the absolute value of the loading matrix}
}
\value{
a 0-1 matrix, indicating the bi-factor structure learned from the loading matrix
}
\description{
compress a bi-factor structure from loading
}
