% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm_cons.R
\name{Bf_cons_pair}
\alias{Bf_cons_pair}
\title{bi-factor constraint pair of columns}
\usage{
Bf_cons_pair(G)
}
\arguments{
\item{G}{the number of group factors}
}
\value{
A list containing all pairs of columns in the equality constraints
}
\description{
bi-factor constraint pair of columns
}
