% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bf_dist}
\alias{bf_dist}
\title{bi-factor approximation evaluation}
\usage{
bf_dist(L)
}
\arguments{
\item{L}{the input loading matrix}
}
\value{
the distance of the given loading matrix to the space of bi-factor loading matrix
}
\description{
bi-factor approximation evaluation
}
