% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{BF_permutation}
\alias{BF_permutation}
\title{Bi-factor permutation}
\usage{
BF_permutation(G)
}
\arguments{
\item{G}{the number of group factors of a bi-factor model}
}
\value{
list of all possible permutation matrix of the group factors
}
\description{
Bi-factor permutation
}
