% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_step.R
\name{bic_process}
\alias{bic_process}
\title{Model Selection via BIC}
\usage{
bic_process(
  G_min,
  G_max,
  S,
  n,
  rho_0 = 1,
  rho_sigma = 10,
  theta = 0.25,
  tol = 0.01,
  max_iter = 100,
  random_starts = 50
)
}
\arguments{
\item{G_min}{the lower bound for the number of factors}

\item{G_max}{the upper bound for the number of factors}

\item{S}{the sample covariance matrix}

\item{n}{the sample size}

\item{rho_0}{the initial value of rho in ALM method, default is 1}

\item{rho_sigma}{the hyper-parameter in ALM method, default is 10}

\item{theta}{the hyper-parameter in ALM method, default is 0.25}

\item{tol}{the tolerance for the ALM method, default is 0.01}

\item{max_iter}{the maximal iteration number, default is 100}

\item{random_starts}{the number of random starting points, default is 50}
}
\value{
A list containing:
\itemize{
\item \code{G_est}: the selected number of group factors based on BIC
\item \code{L_est}: the estimated loading matrix given G_est group factors
\item \code{D_est}: the estimated unique variance matrix given G_est group factors
\item \code{Psi_est}: the estimated correlation matrix given G_est group factors
\item \code{Q_est}: the estimated bi-factor structure given G_est group factors
}
}
\description{
Model Selection via BIC
}
