% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{Compare_Q}
\alias{Compare_Q}
\title{Calculate the average intersection of two bi-factor structure}
\usage{
Compare_Q(Q, Q_est)
}
\arguments{
\item{Q}{the default bi-factor structure}

\item{Q_est}{the estimated bi-factor structure}
}
\value{
the average intersection of two bi-factor structure ranges between 0 to 1
}
\description{
Calculate the average intersection of two bi-factor structure
}
