% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm_cons.R
\name{cons_gd}
\alias{cons_gd}
\title{bi-factor constraint gradient function}
\usage{
cons_gd(L, J, G, gamma, rho, Pair)
}
\arguments{
\item{L}{the loading matrix}

\item{J}{the number of items}

\item{G}{the number of group factors}

\item{gamma}{a matrix-like ALM parameter}

\item{rho}{a scalar-like ALM parameter}

\item{Pair}{a list containing all pairs of columns in the equality constraints}
}
\value{
a vector, which is the gradient of the bi-factor constraint value function
}
\description{
bi-factor constraint gradient function
}
