% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_step.R
\name{Ebfa_solver}
\alias{Ebfa_solver}
\title{EBFA solver}
\usage{
Ebfa_solver(
  S,
  G,
  rho_0 = 1,
  rho_sigma = 10,
  theta = 0.25,
  tol = 0.01,
  max_iter = 100,
  random_starts = 50
)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{G}{the number of group factors}

\item{rho_0}{the initial value of rho in ALM method, default is 1}

\item{rho_sigma}{the hyper-parameter in ALM method, default is 10}

\item{theta}{the hyper-parameter in ALM method, default is 0.25}

\item{tol}{the tolerance for the ALM method, default is 0.01}

\item{max_iter}{the maximal iteration number, default is 100}

\item{random_starts}{the number of random starting points, default is 50}
}
\value{
A list containing:
\itemize{
\item \code{nll_est}: the negative-log-likelihood of the estimated parameters of the bi-factor model
\item \code{L_est}: the estimated loading matrix
\item \code{D_est}: the estimated unique variance matrix
\item \code{Psi_est}: the estimated correlation matrix
\item \code{Q_est}: the estimated bi-factor structure
}
}
\description{
EBFA solver
}
