% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{Exact_matching}
\alias{Exact_matching}
\title{Exact Match Criterion for bi-factor structure}
\usage{
Exact_matching(Q, Q_est, Rot_list = list())
}
\arguments{
\item{Q}{the default bi-factor structure}

\item{Q_est}{the estimated bi-factor structure}

\item{Rot_list}{the rotation matrix considered in the criterion, default is an empty list}
}
\value{
a 0-1 variable, which equals to 1 when the bi-factor structure is correctly learned and 0 otherwise
}
\description{
Exact Match Criterion for bi-factor structure
}
