% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gd_rp}
\alias{gd_rp}
\title{chain rule of the gradient of Phi}
\usage{
gd_rp(x, dU, G)
}
\arguments{
\item{x}{the input vector}

\item{dU}{the gradient in the chain rule}

\item{G}{the number of group factors}
}
\value{
a intermediate result for calculating the gradient of the ALM method
}
\description{
chain rule of the gradient of Phi
}
