% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_alm}
\alias{init_alm}
\title{initial value of bi-factor optimization}
\usage{
init_alm(J, G)
}
\arguments{
\item{J}{the number of items}

\item{G}{the number of group factors}
}
\value{
the initial value of bi-factor optimization
}
\description{
initial value of bi-factor optimization
}
