% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{MSE_parameter}
\alias{MSE_parameter}
\title{Calculate the mean squared error of bi-factor parameters after rotations}
\usage{
MSE_parameter(plist, tlist, Rot_list = list())
}
\arguments{
\item{plist}{A list containing:
\itemize{
\item \code{L_est}: the estimated loading matrix
\item \code{D_est}: the estimated unique variance matrix
\item \code{Psi_est}: the estimated correlation matrix
}}

\item{tlist}{A list containing:
\itemize{
\item \code{L_true}: the true loading matrix
\item \code{D_true}: the true unique variance matrix
\item \code{Psi_true}: the true correlation matrix
}}

\item{Rot_list}{the rotation list for evaluation}
}
\value{
A list containing:
\itemize{
\item \code{MSE_L}: the MSE of loading matrix
\item \code{MSE_D}: the MSE of unique variance matrix
\item \code{MSE_Psi}: the MSE of correlation matrix
}
}
\description{
Calculate the mean squared error of bi-factor parameters after rotations
}
