% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nll}
\alias{nll}
\title{negative-log-likelihood function}
\usage{
nll(L, Phi, D, S)
}
\arguments{
\item{L}{input of loading matrix}

\item{Phi}{input of correlation matrix}

\item{D}{input of diagonal matrix}

\item{S}{sample covariance matrix}
}
\value{
negative-log-likelihood of data
}
\description{
negative-log-likelihood function
}
