% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{para_decompose}
\alias{para_decompose}
\title{transformation of free parameters into the form of loading matrix, correlation matrix and unique variance matrix}
\usage{
para_decompose(x, J, G)
}
\arguments{
\item{x}{input of free parameters of the full bi-factor model}

\item{J}{the number of items}

\item{G}{the number of group factors}
}
\value{
A list containing:
\itemize{
\item \code{L}: the loading matrix
\item \code{D}: the unique variance matrix
\item \code{d}: the parameter of input satisfying diag(d^2) = D
\item \code{Phi}: the Cholesky decomposition of the correlation matrix
\item \code{Psi}: the correlation matrix
\item \code{Cov}: the true covariance matrix
}
}
\description{
transformation of free parameters into the form of loading matrix, correlation matrix and unique variance matrix
}
