% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{parameter_generator}
\alias{parameter_generator}
\title{generate the loading matrix, correlation matrix ,unique variance matrix and covariance matrix given factor structure}
\usage{
parameter_generator(Q, case = "exact")
}
\arguments{
\item{Q}{the factor structure matrix}

\item{case}{the case for generating the parameters, 'exact' indicates an exact bi-factor structure and 'approx' indicates an approximate bi-factor structure}
}
\value{
A list containing:
\itemize{
\item \code{L}: the true loading matrix
\item \code{Psi}: the true correlation matrix
\item \code{D}: the true unique variance matrix
\item \code{Cov}: the true covariance matrix
}
}
\description{
generate the loading matrix, correlation matrix ,unique variance matrix and covariance matrix given factor structure
}
