% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{Q_generator}
\alias{Q_generator}
\title{generate a bi-factor structure as an example}
\usage{
Q_generator(J, G)
}
\arguments{
\item{J}{the number of item}

\item{G}{the number of factors}
}
\value{
a 0-1 matrix indicating the bi-factor structure
}
\description{
generate a bi-factor structure as an example
}
