% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{sampling_process}
\alias{sampling_process}
\title{generate the sample covariance matrix given the covariance matrix}
\usage{
sampling_process(Cov, n)
}
\arguments{
\item{Cov}{the covariance matrix}

\item{n}{the sample size}
}
\value{
the sample covariance matrix
}
\description{
generate the sample covariance matrix given the covariance matrix
}
