% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stan_trans}
\alias{stan_trans}
\title{re-parameterization of Phi}
\usage{
stan_trans(x, G)
}
\arguments{
\item{x}{input of free parameters in the correlation matrix}

\item{G}{the number of group factors}
}
\value{
the Cholesky decomposition of the correlation matrix
}
\description{
re-parameterization of Phi
}
